Use ints for output_windings and sums
We don't expect the winding number to exceed 2 billion here. In theory it could since you can get a self-overlapping triangle with 6 billion vertices if your size_t is long enough. It's kind of a construed case though.

+++ src/polygon_algorithms/Contains.cl
@@ -24,7 +24,7 @@ R"kernel(
  * \param sums Some scratch space to store the computed winding numbers before
  * summing them all up within one work group.
  */
-void kernel contains(global const int2* input_data_points, const long total_vertices, const int2 point, const int include_edges, global long* output_windings, local long* sums) {
+void kernel contains(global const int2* input_data_points, const long total_vertices, const int2 point, const int include_edges, global int* output_windings, local int* sums) {
 	//Compute the winding number contributed by one line segment.
 	const uint global_id = get_global_id(0);
 	const int2 previous = input_data_points[global_id];
